First version of Functions that print Char and String
