+++ main.c
@@ -1,0 +1,8 @@
+#include "holberton.h"
+
+int main(void)
+{
+	_printf("Hello my name is: %s and %c", "erik", 'c');
+	return (0);
+}
+
